requirements.txt
dspy-ai>=2.6.0
streamlit>=1.40.0
pandas>=2.2.0
